docs, move Bvh2Node into its own file
